Migrate to JUnit 5 Jupiter (with Mockito migration).
This follows the migration steps described in the docs under
'Migrating from JUnit 4'.

During the migration a few things had to be corrected due to changes in
behaviour and effects from deprecations.

The first is that lenient has to be declared on all mocks that wont be
used by all tests. This is due to MockitoJUNITRunner verifying that a
mock is used by atleast one test whereas MockitoSettings, with strict
option on, verifies that all test uses a mock.

The second is that the test 'nonArenaSignNoFun' no longer reports a
false positive after upgrading Mockito. This can be verified by
debugging the test before this commit.